Former-commit-id: 13bf51a07ae12a71ce2498c7ce07e48e11ff0517

+++ Sources/Internal/Base/IntrospectionBase.cpp
@@ -44,7 +44,7 @@ InspDesc::InspDesc(const char *_text, const EnumMap* _enumMap)
 	, enumMap(_enumMap)
 { }
 
-InspMember::InspMember(const char *_name, const InspDesc &_desc, const int _offset, const MetaInfo *_type, int _flags /* = 0 */)
+InspMember::InspMember(const char *_name, const InspDesc &_desc, const long int _offset, const MetaInfo *_type, int _flags /* = 0 */)
 	: name(_name), desc(_desc), offset(_offset), type(_type), flags(_flags)
 { }
